+++ lib/igdb_client.rb
@@ -4,10 +4,26 @@ require "faraday"
 require "active_support/all"
 
 module IgdbClient
-  autoload :ApiClient, "igdb_client/api_client"
+  autoload :Api,      "igdb_client/api"
+  autoload :Endpoint, "igdb_client/endpoint"
+  autoload :Request,  "igdb_client/request"
+
   module Constants
     autoload :Endpoints, "igdb_client/constants/endpoints"
   end
+
+  module Query
+    autoload :Builder, "igdb_client/query/builder"
+
+    module Fields
+      autoload :Base,   "igdb_client/query/fields/base"
+      autoload :Field,  "igdb_client/query/fields/field"
+      autoload :Id,     "igdb_client/query/fields/id"
+      autoload :Limit,  "igdb_client/query/fields/limit"
+      autoload :Search, "igdb_client/query/fields/search"
+    end
+  end
+
   module Twitch
     autoload :OauthClient, "igdb_client/twitch/oauth_client"
   end